Major update
Added variable aliasing for de-identification. So now if a variable appears in different files under different names (e.g., "EncounterCSN" and "csn"), it will be de-identified using the same de-identification map.

Added if-else statement to determine which OMOP data set to use, the one with person IDs or the one with patient keys.