Update script template
Replaced OS-specific paths for a relative path approach. This requires the user to specify the directory depth in the directory tree structure of relevant directories, i.e., project directory, IRB directory, and the IDR data requests directory. For this a new function w
as implemented in DRAPI, "successiveParents".

Added test script for the function "successiveParents."