Add string handling for ID variables
Major update: add string handling for ID variables. This was implemented in the following files
  - "deIdentify.py"
  - "makeMapsFromOthers.py"
part of this relies on some metadata constants which are stored in `drapi.constants` in the "constants" and "phiVariables" modules. It might be a good idea to start standardize the de-identification process for different data sources like
  - BO
  - Notes
  - OMOP
  - Zip Codes

Added multiple format handling for `makeMaps`. Now de-identification maps can be made with different de-identified ID formats and column headers.

Updated meta-variable and path variable construction to more generalized methods, e.g., using `successiveParents`.

Organized variable layout.

Replaced `logging.debug` with `logging.info` for easier use in the interpreter. Currently if the log level is set to `DEBUG`, then the IPython interpreter prints out debugging information when the user uses tab-key for completion.