Update templates
OS-specific path construction updates
- Added paths
- Added warning message and exit statement
- Removed unused variables

Note that these changes are conspicuously absent from "concatenateMaps.py". These changes may be unnecessary because of the use of the `pathlib` class `Path`, which might automatically detect the OS. If so, then the OS-specific path construction code blocks can be removed from all the scripts.